annoying warning
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@12821 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/checksum/ipv4.cpp
@@ -107,7 +107,7 @@ RTDECL(bool) RTNetIPv4IsHdrValid(PCRTNETIPV4 pIpHdr, size_t cbHdrMax, size_t cbP
         return false;
     if (RT_UNLIKELY((size_t)pIpHdr->ip_hl * 4 > cbHdrMax))
     {
-        Assert(pIpHdr->ip_hl * 4 > cbPktMax); /* You'll hit this if you mapped/copy too little of the header! */
+        Assert((size_t)pIpHdr->ip_hl * 4 > cbPktMax); /* You'll hit this if you mapped/copy too little of the header! */
         return false;
     }
     if (RT_UNLIKELY(pIpHdr->ip_v != 4))
